ADD: Added a basic proof-of-concept key exchange project that simulates a key exchange by encoding an ECC public key Q into an array of bytes that can be sent and then decoding it back into a public key Q. The project shows the X and Y parameters for Q so you can validate that the "key exchange" worked
TODO: Add a sign and verify function for the key exchange as the ECIES ECC implementation allows for two shared information S1 and S2 which can be used to sign and verify the public keys.

TODO: Add a key exchange and block cipher test function to test key exchange and encrypt/decrypt

NOTE: If shared information S1 and S2 is provided by the users it is possible to sign the publickeys and stop man-in-the-middle attacks, you CANNOT do this with the normal ECDH key exchange
protocol.