Accidentally forgot to mention in last commit
ADD: Added my own custom sign/verifiable public key exchange for ECC when using the ECIES implementation. Because ECIES is being used public key are signed and verified using the shared information S1 if initiating the key exchange or S2 if responding to a key exchange.

It's not as eloquent as Diffie-Hellman exchange, but it works, please read the notes below as well.

NOTE: Since ECIES is being used,the two shared information values S1 and S2 makes it possible to sign and verify public keys. If both users specify two secure (non-trivial or blank) shared information variables S1 and S2 then it makes a man-in-the-middle attack VERY difficult, signing public keys is something ECHD does not support, this is my own sign/verify scheme, lets call it ECG exchange.

NOTE: Users can choose not to provide two shared information (shared passphrases) values, but it makes it impossible to sign/verify the public keys.

NOTE: The key exchange scheme is flawed if only one S value is provided (S1 or S2) as you could determine that the S value is null because the hash would pass without an S1 or S2 given. Therefore a user MUST specify an S1 and S2 OR not specify an S value AT ALL. I would like to make S1 and S2 mandatory.

NOTE: To maintain 256-bit security level for our project each passphrase (S1/S2) can and SHOULD be 32 characters long.