ADD: Added the Burrows Wheeler Transform implementation
The BWT implementation is an optimized version of the BWT transforms
which does not need an entire n x n matrix to perform that transform
instead it uses a suffix array which makes it much faster. The BWT
class used is based off the source of an implementation found in
the kanzi project.