MODIFIED: book_path and shared_info default values
Changed the way that book_path and shared_info retrieve the default value. Rather then storing the default value into the database now it is hard coded into DBAccessor. This is to prevent a second query if the user has the default values for these. Also, this allows for book path and shared info tables to be upgraded to use foreign keys.