ADDED: a text box watcher
Updated the message sending classes to now watch the text box so the number of characters can be kept track of. This will be further extended to show the user how many text messages they are going to be sending. It will have to take into account whether the contact is trusted or not and whether steganography is being used.