ADDED: KeyExchange thread
This has not yet been tested but the thread currently just runs through the differnt numbers (first trusted contacts that are being removed from the trusted list, and second the untrusted contacts that need a key exchange). The key exchange still needs to be implemented with sms and the receiver needs to be set up to accept key exchange messages and handle them properly.