FIXED: new numbers for EditNumber
AddContact now uses EditNumber to add new numbers to the contact. Also, fixed a bug that Shared Info and book paths would not get update to the database. Finally, found a bug in ManageContactsActivity where click on one of the contact's children (numbers when expanded) will use the call back onItemLongClickListener given a 'bad' position therefore mis-using the AddContact activity.