FIXED: a bug with updating shared info and books
Updating shared info and book paths now works correctly so that if only 1 of the two is deleted then it will still default back to the default value. If both are deleted then both are defaulted back to the default value. Also, found and fixed a bug that the first number of a contact would be treated as a new number in editNumber.