DELETED: prephase2
This project has long been done. This project implemented a more
spofisticated database that relyed less on the native database, numbers
that were not found in tinfoil-sms's database were able to be sent/receive
messages to the user through tinfoil-sms, the UI had some improvements as
well as more features were added such as remove contacts, send a message
to a new contact/number.