Fixed a bug in the key exchange on verify for shared info (S)
Fixed a bug in the key exchange where the verifyPubKey() was using the
opposite shared information (S2 instead of S1) to verify the signed public key.
If you are RECEIVING a signed publickey from the initiator of the exchange
the signature is verified using S1, if you are receiving a signed publickey
back from the respondent the signature is verified using S2.