MODIFIED: the layout and the logic of sendSMS
Since the queue was implemented contacts now are required to be in the database to be able to have messages sent to them. Therefore rather then asking if the user would like to add the contact a message send the the contact will automatically add them to the database (if they are not already added. Also, fixed a few bugs in the layout so now both edittexts will stay on screen even with the popup of the onscreen keyboard. Finally, the boarder of both edittexts has been changed back to white, this is still open to change.