MODIFIED: key exchange
Key exchange now will add a message to the message queue to send out a key exchange message. The content of the key exchange message is not correct right now and will be replaced as soon as the encrpytion library is added to the project. Also, The key exchange message shows up in the contact's conversation for now but again once the encrpytion library is added it will be properly handled.