FIXED: key exchange bug
Mixed up the initiator flag. The Number's initiator flag must be set to true when the USER is the initiator. If the contact is the initiator then the flag is set to false. It is slightly confusing but in order for the flag to work correclty it has to be about the user (and therefore the inverse of what the contact is).