ADDED: Number db query
Added the number db query, this can be used when the rest of the contact's information (Such as other numbers, or their name or the messages from that contact) are not needed. Rather then getting the entire TrustedContact object from the db a query and then accessing only a single Number stored.