ADDED: the crypto library to the libs
Android allows for librarys to automatically be added to the build path if placed in the libs folder. Also, re-added the user key generation since it can now use the library. The reason why it has to be in the libs folder and not added to the build path through eclipse is that the library will not be accessible at run time.