Fixed encryption issue, had to use Base64, nonce support working too
Fixed the encryption issue, this solution was to use Base64 encoding
as the message was being corrupted by SMS transmission protocol when
being sent as the raw bytes. The issue with base64 is it has an
overhead of about ~33% to the length of the message.

Currently, nonce support is also working very well, which is being used
with the block cipher in CTR mode.