FIXED: a bug with the auto complete
Fixed a bug with the autocomplete getting a null pointer exection. Also, updated the options menu and removed some of the unneeded items. Also, condensed the encryption and key exchange options into 1 category and compressed all the items about the public key into 1 item which will be an activity to facilitate that.