ADDED: public key view activity
Added an activity that allows the user to view their own public key. This activity will be extended to allow the user to export their public key to a file to allow for a non-sms key exchange to take place. Also, adjusted Edit number to have a subheading for the list of radio buttons.