Completed support for sending the message counter used by the nonce
The message counter is now sent unencrypted prefixed to the start of
the encrypted message. The message counter is used by the cryptographic
nonce to determine the nonce in the sequence of nonces generated
to use. It can also be used to correctly order messages when they
are recieved, if they are not properly transmitted.