ADDED: support for default sms app handling
Tinfoil-SMS will now show up as a possible default app when a user attempts to send a sms from a non sms app. Simple case would be clicking the message button while in the phone app. If no sms default app is defined then the list of apps will show and Tinfoil-sms will be included. Also, adjusted the bug reporting to allow for the user to disable pug reporting.