ADDED: new tab layout
Fixed the db method for retrieving contacts from the database by converting numerous single queries into a single query with an inner join. Also, reduced the amount of information retrieved for the queries since most of the info was not actually needed.
Next an option was added to get all the contacts, only the trusted ones, or only the untrusted ones. This was added to support the tab options for managing trusted and untrusted contacts.
Finally added the layout and tabactivity for managing the trusted and untrusted contacts.