FIXED: Manage Contacts trusted contact update
When a key exchange was received the list of trusted/untrusted users is now updated. It is not perfect yet, since it will work once and then the user must exit out of the activity (and then out of the settings list) and then back in order for it to actually be able to work again.