FIXED: initial key exchange message
When the user initial imports a key exchange their shared secrets are not set, if another user initiates a key exchange with them it would have had a null pointer exception. Now the key exchange message will be handled as a manual key exchange message. Once the user clicks onto it they will be prompted to enter the shared secrets.