FIXES: message view thread and notifications.
MessageView loading thread is now a seperate class. Also, fixed the notifications to properly cancel in most cases. However the notifications do not work correctly if you are in message view (active) and the contact sends you a message). The notification will remain on in the notifications bar until dismissed or the screen is turned back on (after being turned off).