FIXED: key exchange direction issue
Since the key exchange is directed, the reciever's reply is based on the fact that they are the receiver. The issue arose from key exchanges that were rejected and then sent from the reciever. This causes both users involved to be expecting a key exchange reply and rejecting any attempts to start key exchanges (because of the possible man in the middle attack). Now the user is given the choose whether to accept the key exchange or not, and is warned about the security vunerability.