FIXED: key exchange issue with user info
Key exchange would fail since the user's information was not in memory. Now before the key exchange message is created the user is checked for whether it is valid and will retrieve it from the database if it is not in memory. If the user's info is not found in the database, (aka has not been created, or more likely has been deleted) an Invalid db state error is thrown.