FIXED: message limit issue
Message limit was built to support only one contact (making use of sqlite's 'min' function). It will now properly update the messages (delete old ones in order to maintain the max number of messages). However, 1 is the lowest message limit (0 is still accepted but is treated as 1). Finally, messages are not 'actively' deleted, which means if the threshold is lower then the previous threshold there will be no apparent changes (the oldest message will be deleted and replaced by the newest one).