REFACTORED: change log layout
Updated the play store listings to better mirror the new layout of Google's Play store listing. Since the change log is now attached with the apk and not the listing they are stored seperately. It will also be used to keep a more visible history of changes for the translations.