FIXED: send new key exchange to new contact.
Sending a key exchange from the 'new key exchange' activity did not have a check to ensure that the contact was in the database. Thus any number entered that was not already a contact on the phone would silently fail. Now the contact is added so the message will be sent (same as composing a message to a new contact).